Introduction to Neuroscience
This lesson is a foundational introduction to Neuroscience, with emphasis on anatomy of the human brain and function of the nervous system. Topics include cellular components of the nervous system, neuroanatomy and brain organisation, functional systems of the brain, and an introduction to leading research and advancing frontiers in neuroscience.